new implementation for db_graph_get_perfect_path - add functions for node_action and node_condition (more details in following lines)
The functions for printing supernodes, tip clipping, and so on, have two new paramenters with funtions to executed at each node and the condition to be used to skip node.

For example in printing supernodes one might decide to mark all nodes as visited and intiated a new supernode only from "non-visited" nodes. These two actions can be more sophisticated, the user has full control.

+++ include/graph/file_reader.h
@@ -6,14 +6,14 @@
 //these routines return the length of the read sequence, for the binary file is all the kmers conctenated
 
 //for fasta
-int load_fasta_data_from_filename_into_graph(char* filename, long long * bad_reads, int max_read_length, dBGraph* db_graph);
+long long load_fasta_data_from_filename_into_graph(char* filename, long long * bad_reads, int max_read_length, dBGraph* db_graph);
 
 //for fastq
-int load_fastq_data_from_filename_into_graph(char* filename, long long * bad_reads,  char quality_cut_off, int max_read_length, dBGraph* db_graph);
+long long load_fastq_data_from_filename_into_graph(char* filename, long long * bad_reads,  char quality_cut_off, int max_read_length, dBGraph* db_graph);
 
 //for binary
-int load_binary_data_from_filename_into_graph(char* filename, dBGraph* db_graph);
+long long  load_binary_data_from_filename_into_graph(char* filename, dBGraph* db_graph);
 
-int load_binary_data_from_filename_into_graph_with_unique_kmers(char* filename, dBGraph* db_graph);
+long long load_binary_data_from_filename_into_graph_with_unique_kmers(char* filename, dBGraph* db_graph);
 
 #endif /* FILE_READER_H_ */
